PP-15481 Restructure transaction directories
