+++ .secrets.baseline
@@ -278,42 +278,42 @@
         "filename": "src/views/simplified-account/home/all-service-transactions/index.njk",
         "hashed_secret": "8a764e7de8baf91c09a0141c9d0653d6c62300fa",
         "is_verified": false,
-        "line_number": 95
+        "line_number": 42
       },
       {
         "type": "Base64 High Entropy String",
         "filename": "src/views/simplified-account/home/all-service-transactions/index.njk",
         "hashed_secret": "c15bfbed4cf339878d4c69339491df034b592627",
         "is_verified": false,
-        "line_number": 101
+        "line_number": 48
       },
       {
         "type": "Base64 High Entropy String",
         "filename": "src/views/simplified-account/home/all-service-transactions/index.njk",
         "hashed_secret": "a231af987e0851a56413194a38d46620281846c9",
         "is_verified": false,
-        "line_number": 109
+        "line_number": 56
       },
       {
         "type": "Base64 High Entropy String",
         "filename": "src/views/simplified-account/home/all-service-transactions/index.njk",
         "hashed_secret": "d9d993cc5556aa0fa6e18b26fdfd4fa8527db044",
         "is_verified": false,
-        "line_number": 114
+        "line_number": 61
       },
       {
         "type": "Base64 High Entropy String",
         "filename": "src/views/simplified-account/home/all-service-transactions/index.njk",
         "hashed_secret": "ce50854e96f8dd430451450f2e7334e4b48db7f0",
         "is_verified": false,
-        "line_number": 119
+        "line_number": 66
       },
       {
         "type": "Base64 High Entropy String",
         "filename": "src/views/simplified-account/home/all-service-transactions/index.njk",
         "hashed_secret": "34f47d2f3c482850fd7f6de8564a8aa73afe87bc",
         "is_verified": false,
-        "line_number": 124
+        "line_number": 71
       }
     ],
     "src/views/simplified-account/home/all-service-transactions/nosearch.njk": [
@@ -322,86 +322,86 @@
         "filename": "src/views/simplified-account/home/all-service-transactions/nosearch.njk",
         "hashed_secret": "8a764e7de8baf91c09a0141c9d0653d6c62300fa",
         "is_verified": false,
-        "line_number": 80
+        "line_number": 39
       },
       {
         "type": "Base64 High Entropy String",
         "filename": "src/views/simplified-account/home/all-service-transactions/nosearch.njk",
         "hashed_secret": "c15bfbed4cf339878d4c69339491df034b592627",
         "is_verified": false,
-        "line_number": 86
+        "line_number": 45
       },
       {
         "type": "Base64 High Entropy String",
         "filename": "src/views/simplified-account/home/all-service-transactions/nosearch.njk",
         "hashed_secret": "a231af987e0851a56413194a38d46620281846c9",
         "is_verified": false,
-        "line_number": 94
+        "line_number": 53
       },
       {
         "type": "Base64 High Entropy String",
         "filename": "src/views/simplified-account/home/all-service-transactions/nosearch.njk",
         "hashed_secret": "d9d993cc5556aa0fa6e18b26fdfd4fa8527db044",
         "is_verified": false,
-        "line_number": 99
+        "line_number": 58
       },
       {
         "type": "Base64 High Entropy String",
         "filename": "src/views/simplified-account/home/all-service-transactions/nosearch.njk",
         "hashed_secret": "ce50854e96f8dd430451450f2e7334e4b48db7f0",
         "is_verified": false,
-        "line_number": 104
+        "line_number": 63
       },
       {
         "type": "Base64 High Entropy String",
         "filename": "src/views/simplified-account/home/all-service-transactions/nosearch.njk",
         "hashed_secret": "34f47d2f3c482850fd7f6de8564a8aa73afe87bc",
         "is_verified": false,
-        "line_number": 109
+        "line_number": 68
       }
     ],
-    "src/views/simplified-account/transactions/index.njk": [
+    "src/views/simplified-account/services/transactions/list/transaction-index.njk": [
       {
         "type": "Base64 High Entropy String",
-        "filename": "src/views/simplified-account/transactions/index.njk",
+        "filename": "src/views/simplified-account/services/transactions/list/transaction-index.njk",
         "hashed_secret": "8a764e7de8baf91c09a0141c9d0653d6c62300fa",
         "is_verified": false,
-        "line_number": 100
+        "line_number": 46
       },
       {
         "type": "Base64 High Entropy String",
-        "filename": "src/views/simplified-account/transactions/index.njk",
+        "filename": "src/views/simplified-account/services/transactions/list/transaction-index.njk",
         "hashed_secret": "c15bfbed4cf339878d4c69339491df034b592627",
         "is_verified": false,
-        "line_number": 106
+        "line_number": 52
       },
       {
         "type": "Base64 High Entropy String",
-        "filename": "src/views/simplified-account/transactions/index.njk",
+        "filename": "src/views/simplified-account/services/transactions/list/transaction-index.njk",
         "hashed_secret": "a231af987e0851a56413194a38d46620281846c9",
         "is_verified": false,
-        "line_number": 114
+        "line_number": 60
       },
       {
         "type": "Base64 High Entropy String",
-        "filename": "src/views/simplified-account/transactions/index.njk",
+        "filename": "src/views/simplified-account/services/transactions/list/transaction-index.njk",
         "hashed_secret": "d9d993cc5556aa0fa6e18b26fdfd4fa8527db044",
         "is_verified": false,
-        "line_number": 119
+        "line_number": 65
       },
       {
         "type": "Base64 High Entropy String",
-        "filename": "src/views/simplified-account/transactions/index.njk",
+        "filename": "src/views/simplified-account/services/transactions/list/transaction-index.njk",
         "hashed_secret": "ce50854e96f8dd430451450f2e7334e4b48db7f0",
         "is_verified": false,
-        "line_number": 124
+        "line_number": 70
       },
       {
         "type": "Base64 High Entropy String",
-        "filename": "src/views/simplified-account/transactions/index.njk",
+        "filename": "src/views/simplified-account/services/transactions/list/transaction-index.njk",
         "hashed_secret": "34f47d2f3c482850fd7f6de8564a8aa73afe87bc",
         "is_verified": false,
-        "line_number": 129
+        "line_number": 75
       }
     ],
     "src/views/transactions/index.njk": [
@@ -882,5 +882,5 @@
       }
     ]
   },
-  "generated_at": "2026-06-29T16:33:07Z"
+  "generated_at": "2026-07-10T09:51:34Z"
 }
