+++ src/controllers/simplified-account/services/transactions/index.ts
@@ -1,6 +1,6 @@
 import * as list from './transaction-list.controller'
-import * as detail from './detail/transactions-detail.controller'
-import * as refund from './transaction-refund.controller'
+import * as detail from './detail/transaction-detail.controller'
+import * as refund from './refund/transaction-refund.controller'
 import * as downloadCsv from './csv-download/transaction-csv-download.controller'
 
 export { detail, list, refund, downloadCsv }
