+++ src/views/simplified-account/home/all-service-transactions/nosearch.njk
@@ -21,53 +21,12 @@
         {% endif %}
       </div>
 
-      <form method="GET" novalidate autocomplete="off" action="{{ searchLink }}">
-        {% include "simplified-account/transactions/filter.njk" %}
+      {% include 'simplified-account/services/transactions/list/partials/_search-form.njk' %}
 
-
-        {% set advancedFilters %}
-          {% include "simplified-account/transactions/advanced-filter.njk" %}
-        {% endset %}
-
-        {{
-          govukAccordion({
-            id: "accordion-default",
-            classes: "disable-accordion-controls",
-            items: [
-              {
-                heading: {
-                text: "Advanced search"
-              },
-                content: {
-                html: advancedFilters
-              }
-              }
-            ]
-          })
-        }}
-
-        <div class="govuk-button-group">
-          {{
-            govukButton({
-              text: "Search transactions"
-            })
-          }}
-        </div>
-      </form>
       <hr class="govuk-section-break govuk-section-break--m govuk-section-break--visible" />
 
       {% if isBST %}
-        <details class="govuk-details">
-          <summary class="govuk-details__summary">
-            <span class="govuk-details__summary-text"> How times work in CSV downloads </span>
-          </summary>
-          <div class="govuk-details__text govuk-grid-column-two-thirds" style="margin-bottom: 20px">
-            The UK is on British Summertime (BST) but CSV downloads are in UTC (Coordinated Universal Time), which is
-            the same as GMT (Greenwich Mean Time).<br /><br />
-            If you want to download a day’s transactions, you can change your filters so it’s one hour ahead, for
-            example 01:00:00 to 01:00:00 instead of 00:00:00 to 00:00:00.
-          </div>
-        </details>
+        {% include 'simplified-account/services/transactions/list/partials/_bst-details.njk' %}
       {% endif %}
     </div>
   </div>
