+++ src/views/simplified-account/services/transactions/detail/transaction-detail.njk
