+++ src/views/simplified-account/services/transactions/list/macro/_stripe-transaction-row.njk
