+++ src/views/simplified-account/services/transactions/list/partials/_advanced-filter.njk
@@ -1,0 +1,19 @@
+<div class="transactions-filter-simplified-account">
+  <div class="govuk-grid-row">
+    <div class="govuk-grid-column-one-third inputs-less-margin">{% include "./filter-inputs/_name-input.njk" %}</div>
+    <div class="govuk-grid-column-one-third inputs-less-margin"
+      >{% include "./filter-inputs/_last-four-digits-input.njk" %}</div
+    >
+    <div class="govuk-grid-column-one-third inputs-less-margin"
+      >{% include "./filter-inputs/_metadata-input.njk" %}</div
+    >
+  </div>
+
+  <div class="govuk-grid-row">
+    <div class="govuk-grid-column-one-third inputs-less-margin">
+      <div class="govuk-form-group govuk-!-margin-bottom-0">
+        <fieldset class="govuk-fieldset">{% include "./filter-inputs/_card-brand.njk" %}</fieldset>
+      </div>
+    </div>
+  </div>
+</div>
