+++ src/views/simplified-account/services/transactions/list/partials/_bst-details.njk
@@ -1,0 +1,11 @@
+<details class="govuk-details">
+  <summary class="govuk-details__summary">
+    <span class="govuk-details__summary-text"> How times work in CSV downloads </span>
+  </summary>
+  <div class="govuk-details__text govuk-grid-column-two-thirds" style="margin-bottom: 20px">
+    The UK is on British Summertime (BST) but CSV downloads are in UTC (Coordinated Universal Time), which is the same
+    as GMT (Greenwich Mean Time).<br /><br />
+    If you want to download a day’s transactions, you can change your filters so it’s one hour ahead, for example
+    01:00:00 to 01:00:00 instead of 00:00:00 to 00:00:00.
+  </div>
+</details>
