+++ src/views/simplified-account/services/transactions/list/partials/_csv-link.njk
@@ -1,0 +1,11 @@
+{% if (showCsvDownload) %}
+  {{
+    govukButton({
+      text: "Download CSV",
+      classes: "govuk-button--secondary",
+      href: downloadLink
+    })
+  }}
+{% elif (results.transactions.length) %}
+  <p id="csv-download" class="govuk-body">Filter results to download a CSV of transactions</p>
+{% endif %}
