+++ src/views/simplified-account/services/transactions/list/partials/_filter.njk
@@ -1,18 +1,20 @@
 <div class="transactions-filter-simplified-account" style="width:100%">
   <div class="govuk-grid-row govuk-clearfix">
-    <div class="govuk-grid-column-one-third inputs-less-margin">{% include "./partials/_reference-input.njk" %}</div>
-    <div class="govuk-grid-column-one-third inputs-less-margin">{% include "./partials/_email-input.njk" %}</div>
+    <div class="govuk-grid-column-one-third inputs-less-margin"
+      >{% include "./filter-inputs/_reference-input.njk" %}</div
+    >
+    <div class="govuk-grid-column-one-third inputs-less-margin">{% include "./filter-inputs/_email-input.njk" %}</div>
     <div class="govuk-grid-column-one-third">
-      <fieldset class="govuk-fieldset">{% include "./partials/_status-select.njk" %}</fieldset>
+      <fieldset class="govuk-fieldset">{% include "./filter-inputs/_status-select.njk" %}</fieldset>
     </div>
   </div>
   <div class="govuk-grid-row">
     <div class="govuk-grid-column-one-third">
-      <fieldset class="govuk-fieldset">{% include "./partials/_date-filter-select.njk" %}</fieldset>
+      <fieldset class="govuk-fieldset">{% include "./filter-inputs/_date-filter-select.njk" %}</fieldset>
     </div>
     <div class="govuk-grid-column-one-third">
       <div class="govuk-form-group govuk-!-margin-bottom-0">
-        <fieldset class="govuk-fieldset">{% include "./partials/_date-picker.njk" %}</fieldset>
+        <fieldset class="govuk-fieldset">{% include "./filter-inputs/_date-picker.njk" %}</fieldset>
       </div>
     </div>
     <div class="govuk-grid-column-one-third">
@@ -49,7 +51,7 @@
       <br class="transactions-list__br" />
     </div>
     <div class="govuk-grid-column-one-third">
-      <fieldset class="govuk-fieldset">{% include "./partials/_time-picker.njk" %}</fieldset>
+      <fieldset class="govuk-fieldset">{% include "./filter-inputs/_time-picker.njk" %}</fieldset>
     </div>
     <div class="govuk-grid-column-one-third"> <br class="transactions-list__br" /></div>
   </div>
