+++ src/views/simplified-account/services/transactions/list/partials/filter-inputs/_date-picker.njk
